Use correct type for iteration

+++ renew.c
@@ -8,15 +8,15 @@
 // Rearrange cells into new dimensions.
 static int rearrange(struct cgol *instance, size_t ysize, size_t xsize) {
     // Store new value in second bit.
-    for (int i = 0; (i < ysize) && (i < instance->ysize); i++) {
-        for (int j = 0; (j < xsize) && (j < instance->xsize); j++) {
+    for (size_t i = 0; (i < ysize) && (i < instance->ysize); i++) {
+        for (size_t j = 0; (j < xsize) && (j < instance->xsize); j++) {
             MGRID_AT(instance, xsize, i, j) =
                 MRGRID_AT(instance, xsize, i, j)
                 | ((RGRID_AT(instance, i, j)) << 1);
         }
     }
     // Update cells.
-    for (int i = 0; i < ysize*xsize; i++) {
+    for (size_t i = 0; i < ysize*xsize; i++) {
         instance->grid[i] >>= 1;
     }
     return SUCCESS;
@@ -24,7 +24,7 @@ static int rearrange(struct cgol *instance, size_t ysize, size_t xsize) {
 
 // Zero out new area after size increase.
 static int zero(struct cgol *instance, size_t ysize, size_t xsize) {
-    for (int i = instance->ysize*instance->xsize; i < ysize*xsize; i++) {
+    for (size_t i = instance->ysize*instance->xsize; i < ysize*xsize; i++) {
         instance->grid[i] = 0;
     }
     return SUCCESS;
